Refactor summary classes
