+++ app/views/contexts/__init__.py
@@ -2,14 +2,12 @@ from .calculated_summary_context import CalculatedSummaryContext
 from .context import Context
 from .hub_context import HubContext
 from .list_context import ListContext
-from .custom_section_summary_context import CustomSectionSummaryContext
 from .final_summary_context import FinalSummaryContext
 from .section_summary_context import SectionSummaryContext
 
 __all__ = [
     "CalculatedSummaryContext",
     "Context",
-    "CustomSectionSummaryContext",
     "FinalSummaryContext",
     "HubContext",
     "ListContext",
